requirements.txt
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
Django==2.2.4
django-chartjs==1.5.0
djangorestframework==3.10.3
idna==2.8
iso8601==0.1.12
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.1
paho-mqtt==1.4.0
plotly==4.1.0
psutil==5.6.3
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-toolbelt==0.9.1
retrying==1.3.3
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
wiotp-sdk==0.9.5
